feat: refactor + add network intercept
