+++ src/bidiMapper/CommandProcessor.ts
@@ -15,18 +15,12 @@
  * limitations under the License.
  */
 
-import {
-  type BrowsingContext,
-  type Cdp,
-  type Input,
-  Message,
-  type Script,
-  type Session,
-} from '../protocol/protocol.js';
+import {Message, type Session} from '../protocol/protocol.js';
 import {LogType, type LoggerFn} from '../utils/log.js';
 import {EventEmitter} from '../utils/EventEmitter.js';
 import type {ICdpConnection} from '../cdp/cdpConnection.js';
 
+import {type BidiParser, BidiNoOpParser} from './BidiParser.js';
 import {BrowsingContextProcessor} from './domains/context/browsingContextProcessor.js';
 import type {BrowsingContextStorage} from './domains/context/browsingContextStorage.js';
 import type {IEventManager} from './domains/events/EventManager.js';
@@ -36,104 +30,6 @@ import type {RealmStorage} from './domains/script/realmStorage.js';
 type CommandProcessorEvents = {
   response: Promise<OutgoingBidiMessage>;
 };
-
-export interface BidiParser {
-  parseAddPreloadScriptParams(
-    params: object
-  ): Script.AddPreloadScriptParameters;
-  parseRemovePreloadScriptParams(
-    params: object
-  ): Script.RemovePreloadScriptParameters;
-  parseGetRealmsParams(params: object): Script.GetRealmsParameters;
-  parseCallFunctionParams(params: object): Script.CallFunctionParameters;
-  parseEvaluateParams(params: object): Script.EvaluateParameters;
-  parseDisownParams(params: object): Script.DisownParameters;
-  parseSendCommandParams(params: object): Cdp.SendCommandParams;
-  parseGetSessionParams(params: object): Cdp.GetSessionParams;
-  parseSubscribeParams(params: object): Session.SubscriptionRequest;
-  parseNavigateParams(params: object): BrowsingContext.NavigateParameters;
-  parseReloadParams(params: object): BrowsingContext.ReloadParameters;
-  parseGetTreeParams(params: object): BrowsingContext.GetTreeParameters;
-  parseCreateParams(params: object): BrowsingContext.CreateParameters;
-  parseCloseParams(params: object): BrowsingContext.CloseParameters;
-  parseCaptureScreenshotParams(
-    params: object
-  ): BrowsingContext.CaptureScreenshotParameters;
-  parsePrintParams(params: object): BrowsingContext.PrintParameters;
-  parseSetViewportParams(params: object): BrowsingContext.SetViewportParameters;
-  parsePerformActionsParams(params: object): Input.PerformActionsParameters;
-  parseReleaseActionsParams(params: object): Input.ReleaseActionsParameters;
-}
-
-class BidiNoOpParser implements BidiParser {
-  parseAddPreloadScriptParams(
-    params: object
-  ): Script.AddPreloadScriptParameters {
-    return params as Script.AddPreloadScriptParameters;
-  }
-
-  parseRemovePreloadScriptParams(
-    params: object
-  ): Script.RemovePreloadScriptParameters {
-    return params as Script.RemovePreloadScriptParameters;
-  }
-
-  parseGetRealmsParams(params: object): Script.GetRealmsParameters {
-    return params as Script.GetRealmsParameters;
-  }
-  parseCallFunctionParams(params: object): Script.CallFunctionParameters {
-    return params as Script.CallFunctionParameters;
-  }
-  parseEvaluateParams(params: object): Script.EvaluateParameters {
-    return params as Script.EvaluateParameters;
-  }
-  parseDisownParams(params: object): Script.DisownParameters {
-    return params as Script.DisownParameters;
-  }
-  parseSendCommandParams(params: object): Cdp.SendCommandParams {
-    return params as Cdp.SendCommandParams;
-  }
-  parseGetSessionParams(params: object): Cdp.GetSessionParams {
-    return params as Cdp.GetSessionParams;
-  }
-  parseSubscribeParams(params: object): Session.SubscriptionRequest {
-    return params as Session.SubscriptionRequest;
-  }
-  parseNavigateParams(params: object): BrowsingContext.NavigateParameters {
-    return params as BrowsingContext.NavigateParameters;
-  }
-  parseReloadParams(params: object): BrowsingContext.ReloadParameters {
-    return params as BrowsingContext.ReloadParameters;
-  }
-  parseGetTreeParams(params: object): BrowsingContext.GetTreeParameters {
-    return params as BrowsingContext.GetTreeParameters;
-  }
-  parseCreateParams(params: object): BrowsingContext.CreateParameters {
-    return params as BrowsingContext.CreateParameters;
-  }
-  parseCloseParams(params: object): BrowsingContext.CloseParameters {
-    return params as BrowsingContext.CloseParameters;
-  }
-  parseCaptureScreenshotParams(
-    params: object
-  ): BrowsingContext.CaptureScreenshotParameters {
-    return params as BrowsingContext.CaptureScreenshotParameters;
-  }
-  parsePrintParams(params: object): BrowsingContext.PrintParameters {
-    return params as BrowsingContext.PrintParameters;
-  }
-  parsePerformActionsParams(params: object): Input.PerformActionsParameters {
-    return params as Input.PerformActionsParameters;
-  }
-  parseReleaseActionsParams(params: object): Input.ReleaseActionsParameters {
-    return params as Input.ReleaseActionsParameters;
-  }
-  parseSetViewportParams(
-    params: object
-  ): BrowsingContext.SetViewportParameters {
-    return params as BrowsingContext.SetViewportParameters;
-  }
-}
 
 export class CommandProcessor extends EventEmitter<CommandProcessorEvents> {
   #contextProcessor: BrowsingContextProcessor;
@@ -196,19 +92,6 @@ export class CommandProcessor extends EventEmitter<CommandProcessorEvents> {
     commandData: Message.RawCommandRequest
   ): Promise<Message.ResultData> {
     switch (commandData.method) {
-      case 'session.status':
-        return CommandProcessor.#process_session_status();
-      case 'session.subscribe':
-        return this.#process_session_subscribe(
-          this.#parser.parseSubscribeParams(commandData.params),
-          commandData.channel ?? null
-        );
-      case 'session.unsubscribe':
-        return this.#process_session_unsubscribe(
-          this.#parser.parseSubscribeParams(commandData.params),
-          commandData.channel ?? null
-        );
-
       case 'browsingContext.create':
         return this.#contextProcessor.process_browsingContext_create(
           this.#parser.parseCreateParams(commandData.params)
@@ -242,6 +125,29 @@ export class CommandProcessor extends EventEmitter<CommandProcessorEvents> {
           this.#parser.parseSetViewportParams(commandData.params)
         );
 
+      case 'cdp.sendCommand':
+        return this.#contextProcessor.process_cdp_sendCommand(
+          this.#parser.parseSendCommandParams(commandData.params)
+        );
+      case 'cdp.getSession':
+        return this.#contextProcessor.process_cdp_getSession(
+          this.#parser.parseGetSessionParams(commandData.params)
+        );
+
+      case 'input.performActions':
+        return this.#contextProcessor.process_input_performActions(
+          this.#parser.parsePerformActionsParams(commandData.params)
+        );
+      case 'input.releaseActions':
+        return this.#contextProcessor.process_input_releaseActions(
+          this.#parser.parseReleaseActionsParams(commandData.params)
+        );
+
+      case 'network.addIntercept':
+        return this.#contextProcessor.process_network_addIntercept(
+          this.#parser.parseAddInterceptParams(commandData.params)
+        );
+
       case 'script.addPreloadScript':
         return this.#contextProcessor.process_script_addPreloadScript(
           this.#parser.parseAddPreloadScriptParams(commandData.params)
@@ -267,22 +173,17 @@ export class CommandProcessor extends EventEmitter<CommandProcessorEvents> {
           this.#parser.parseDisownParams(commandData.params)
         );
 
-      case 'input.performActions':
-        return this.#contextProcessor.process_input_performActions(
-          this.#parser.parsePerformActionsParams(commandData.params)
+      case 'session.status':
+        return CommandProcessor.#process_session_status();
+      case 'session.subscribe':
+        return this.#process_session_subscribe(
+          this.#parser.parseSubscribeParams(commandData.params),
+          commandData.channel ?? null
         );
-      case 'input.releaseActions':
-        return this.#contextProcessor.process_input_releaseActions(
-          this.#parser.parseReleaseActionsParams(commandData.params)
-        );
-
-      case 'cdp.sendCommand':
-        return this.#contextProcessor.process_cdp_sendCommand(
-          this.#parser.parseSendCommandParams(commandData.params)
-        );
-      case 'cdp.getSession':
-        return this.#contextProcessor.process_cdp_getSession(
-          this.#parser.parseGetSessionParams(commandData.params)
+      case 'session.unsubscribe':
+        return this.#process_session_unsubscribe(
+          this.#parser.parseSubscribeParams(commandData.params),
+          commandData.channel ?? null
         );
     }
