+++ src/bidiMapper/bidiMapper.ts
@@ -29,4 +29,4 @@ export type {ICdpClient, ICdpClient as CdpClient} from '../cdp/cdpClient.js';
 export {EventEmitter} from '../utils/EventEmitter.js';
 export type {BidiTransport} from './BidiTransport.js';
 export {OutgoingBidiMessage} from './OutgoingBidiMessage.js';
-export type {BidiParser} from './CommandProcessor.js';
+export type {BidiParser} from './BidiParser.js';
